+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added `spicyweb\embeddedassets\Variable::create()` (`craft.embeddedAssets.create()`) for creating an `EmbeddedAsset` model from user-provided data
 - Added a console command for refreshing embedded asset data by provider
 
+### Changed
+- The console command for refreshing all embedded asset data now accepts both `--volume` and `--provider` options
+
 ## 2.9.1 - 2021-11-30
 
 ### Fixed